Add fire, harmonise ints, move type to own file

+++ src/engine/core.c
@@ -9,9 +9,6 @@ void init_console(void)
     SMS_initSprites();
     SMS_setSpriteMode(SPRITEMODE_NORMAL);  // Set this to the expected format
     SMS_useFirstHalfTilesforSprites(true); // Are sprite memory reads limited?
-    // SMS_loadPSGaidencompressedTiles(blank_tiles_psgcompr, 0);
-    // SMS_loadBGPalette();
-    // SMS_loadSpritePalette();
 
     // Should probably wipe the tile library as well?
     clear_tilemap(0);
@@ -25,7 +22,7 @@ void wait_for_frame(void)
     SMS_copySpritestoSAT(); // Update any sprite shenanigans
 }
 
-void clear_tilemap(unsigned int tile) // Must be an int, not a uint_8t or such
+void clear_tilemap(uint16_t tile) // Must be an int, not a uint_8t or such
 {
     // From sverx, a more efficient method
     SMS_VRAMmemsetW(XYtoADDR(0, 0), tile, SCREEN_ROWS * SCREEN_COLUMNS * 2);
